+++ core/templates/dev/head/pages/learner-dashboard-page/suggestion-modal/suggestion-modal-for-learner-dashboard.service.ts
@@ -17,7 +17,7 @@
  */
 
 require('domain/utilities/UrlInterpolationService.ts');
-require('pages/suggestion_editor/SuggestionModalService.ts');
+require('services/SuggestionModalService.ts');
 
 oppia.factory('SuggestionModalForLearnerDashboardService', [
   '$rootScope', '$uibModal', 'UrlInterpolationService',
