+++ core/templates/dev/head/services/SuggestionModalService.ts
